feat: CLI improvements
- warn when multiple versions of zenstack packages are detected
- check for new version
