+++ packages/schema/src/cli/actions/index.ts
@@ -1,0 +1,3 @@
+export * from './generate';
+export * from './info';
+export * from './init';
