+++ packages/schema/src/cli/index.ts
@@ -3,33 +3,20 @@ import { ZModelLanguageMetaData } from '@zenstackhq/language/module';
 import colors from 'colors';
 import { Command, Option } from 'commander';
 import fs from 'fs';
-import * as semver from 'semver';
 import telemetry from '../telemetry';
-import { PackageManagers } from '../utils/pkg-utils';
 import { getVersion } from '../utils/version-utils';
-import { CliError } from './cli-error';
-import { dumpInfo, initProject, runPlugins } from './cli-util';
+import * as actions from './actions';
 import { loadConfig } from './config';
-
-// required minimal version of Prisma
-export const requiredPrismaVersion = '4.0.0';
 
 const DEFAULT_CONFIG_FILE = 'zenstack.config.json';
 
-export const initAction = async (
-    projectPath: string,
-    options: {
-        prisma: string | undefined;
-        packageManager: PackageManagers | undefined;
-        tag?: string;
-    }
-): Promise<void> => {
+export const initAction = async (projectPath: string, options: Parameters<typeof actions.init>[1]): Promise<void> => {
     await telemetry.trackSpan(
         'cli:command:start',
         'cli:command:complete',
         'cli:command:error',
         { command: 'init' },
-        () => initProject(projectPath, options.prisma, options.packageManager, options.tag)
+        () => actions.init(projectPath, options)
     );
 };
 
@@ -39,46 +26,18 @@ export const infoAction = async (projectPath: string): Promise<void> => {
         'cli:command:complete',
         'cli:command:error',
         { command: 'info' },
-        () => dumpInfo(projectPath)
+        () => actions.info(projectPath)
     );
 };
 
-export const generateAction = async (options: {
-    schema: string;
-    packageManager: PackageManagers | undefined;
-    dependencyCheck: boolean;
-}): Promise<void> => {
-    if (options.dependencyCheck) {
-        checkRequiredPackage('prisma', requiredPrismaVersion);
-        checkRequiredPackage('@prisma/client', requiredPrismaVersion);
-    }
+export const generateAction = async (options: Parameters<typeof actions.generate>[1]): Promise<void> => {
     await telemetry.trackSpan(
         'cli:command:start',
         'cli:command:complete',
         'cli:command:error',
         { command: 'generate' },
-        () => runPlugins(options)
+        () => actions.generate(process.cwd(), options)
     );
-};
-
-const checkRequiredPackage = (packageName: string, minVersion?: string) => {
-    let packageVersion: string;
-    try {
-        // eslint-disable-next-line @typescript-eslint/no-var-requires
-        packageVersion = require(`${packageName}/package.json`).version;
-    } catch (error) {
-        console.error(colors.red(`${packageName} not found, please install it`));
-        throw new CliError(`${packageName} not found`);
-    }
-
-    if (minVersion && semver.lt(packageVersion, minVersion)) {
-        console.error(
-            colors.red(
-                `${packageName} needs to be above ${minVersion}, the installed version is ${packageVersion}, please upgrade it`
-            )
-        );
-        throw new CliError(`${packageName} version is too low`);
-    }
 };
 
 export function createProgram() {
@@ -102,13 +61,12 @@ export function createProgram() {
     );
 
     const configOption = new Option('-c, --config [file]', 'config file');
-
     const pmOption = new Option('-p, --package-manager <pm>', 'package manager to use').choices([
         'npm',
         'yarn',
         'pnpm',
     ]);
-
+    const noVersionCheckOption = new Option('--no-version-check', 'do not check for new version');
     const noDependencyCheck = new Option('--no-dependency-check', 'do not check if dependencies are installed');
 
     program
@@ -124,6 +82,7 @@ export function createProgram() {
         .addOption(pmOption)
         .addOption(new Option('--prisma <file>', 'location of Prisma schema file to bootstrap from'))
         .addOption(new Option('--tag [tag]', 'the NPM package tag to use when installing dependencies'))
+        .addOption(noVersionCheckOption)
         .argument('[path]', 'project path', '.')
         .action(initAction);
 
@@ -133,6 +92,7 @@ export function createProgram() {
         .addOption(schemaOption)
         .addOption(configOption)
         .addOption(pmOption)
+        .addOption(noVersionCheckOption)
         .addOption(noDependencyCheck)
         .action(generateAction);
